chore: 패키지 이동

+++ app/src/main/java/woowacourse/shopping/domain/entity/CartProduct.kt
@@ -9,7 +9,7 @@ data class CartProduct(
         require(count >= MIN_COUNT_LIMIT) { "카트의 수량은 1개 이상이어야 합니다." }
     }
 
-    val totalPrice get(): Int = product.price * count
+    val totalPrice get(): Long = product.price.toLong() * count
 
     fun increaseCount(amount: Int = 1): CartProduct = CartProduct(product, count + amount)
